dogApi/requirements.txt
fastapi
uvicorn
torch
torchvision
faiss-cpu
pillow
requests
numpy
